requirements.txt
faiss-cpu==1.9.0
PyPDF2==3.0.1
openai==1.54.3
python-dotenv==1.0.1
html2text==2024.2.26
python-docx==1.1.2
python-pptx==1.0.2
llama-index==0.11.23
gradio==5.5.0
